[Bug 1282] nav speed is now a double for viewpoints and viewpoints works again
Former-commit-id: 3fec1adac6782cf52cedba9b5ceefe5ad6de7c4a

+++ src/apps/conductor/ViewLocPane.h
@@ -146,7 +146,7 @@ protected:
     int cId, cIso_value, cSc, cMin;
     std::string dataValueName;
 
-    std::vector< long > commandInputs;
+    std::vector< double > commandInputs;
     wxScrolledWindow* scrollWindow;
 
     wxComboBox* _activeflySel;
